pyproject.toml
[tool.poetry]
name = "yara-grep"
version = "0.4.0"
description = "CLI tool for rapid Yara rule prototyping"
authors = ["Aled Mason <aledmason@microsoft.com>"]
readme = "README.md"
repository = "https://github.com/0xa13d/yara-grep"

[tool.poetry.dependencies]
python = "^3.10"
yara-python = "^4.5.1"
tabulate = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ygrep = "yara_grep.cli:main"